+++ CompatibilityLib/common/src/main/java/com/elmakers/mine/bukkit/utility/platform/SkinUtils.java
@@ -3,6 +3,7 @@ package com.elmakers.mine.bukkit.utility.platform;
 import java.util.UUID;
 
 import org.bukkit.entity.Player;
+import org.bukkit.profile.PlayerProfile;
 
 import com.elmakers.mine.bukkit.utility.ProfileCallback;
 
@@ -14,4 +15,6 @@ public interface SkinUtils {
     void fetchProfile(String playerName, ProfileCallback callback);
 
     void fetchProfile(UUID uuid, ProfileCallback callback);
+
+    String toDisguiseFormat(PlayerProfile profile);
 }